fix: stop rounding the ideal-burndown line, causing false plateaus
expected was rounded to the nearest integer per day, which plateaus
for 1+ days whenever total < the release's day span (the common
case) — e.g. 3 test cases over 10 days rounded to [3,3,2,2,2,2,1,1,1,0,0]
instead of a smooth line. It's a continuous reference trend, not a
discrete count like `remaining`, so it's left un-rounded server-side;
the tooltip rounds only for display.

+++ backend/schemas/releases.py
@@ -84,4 +84,4 @@ class ExecutionHistoryItem(BaseModel):
 class BurndownPoint(BaseModel):
     date: date
     remaining: int
-    expected: int | None
+    expected: float | None
